requirements.txt
Django==5.0.7
djangorestframework==3.15.2

django-environ==0.11.2
django-model-utils==4.5.1
django-admin-timestamps==1.2
django-dump-load-utf8==0.0.4

psycopg2-binary==2.9.9